Proof-of-concept: Classed component
This was an idea I had while thinking about how to make Styled Components play nice with Basscss, and while it sounds like Styled Components may get support for a class presetting mechanism built-in, this might be a good approach in the mean time.

It allows for supplying the kinds of object interpolations you usually do to ClassNames, but within a callback function you supply. Alternately, you can just supply strings and such.